requirements.txt
Flask==2.1.0
Flask-RESTx==0.5.1
Flask-SQLAlchemy==2.5.0
Flask-Celery-Helper==1.0.0
Flask-Migrate==3.1.0
Flask-Swagger-UI==3.36.0
redis==3.5.3
celery==5.2.2
python-dotenv==0.19.1
pymysql==1.0.2
pytz==2021.3